services/retriever-py/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
sentence-transformers==3.0.1
faiss-cpu==1.12.0
scikit-learn==1.5.1
numpy==1.26.4
tqdm==4.66.5